pyproject.toml
[project]
name = "nerve"
version = "0.1.0"
description = "Personal AI assistant — lightweight, single-process, purpose-built"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "aiosqlite>=0.21.0",
    "pyyaml>=6.0",
    "python-telegram-bot>=21.0",
    "claude-agent-sdk>=0.1.30",
    "apscheduler>=3.11.0",
    "pyjwt>=2.10.0",
    "bcrypt>=4.2.0",
    "httpx>=0.28.0",
    "websockets>=14.0",
    "click>=8.1.0",
    "telethon>=1.38.0",
    "anthropic>=0.45.0",  # used by memu_bridge for event date resolution
    "memu-py==1.4.0",  # pinned — nerve monkey-patches memU internals (see memu_bridge._patch_sqlite_bugs)
    "watchfiles>=1.0.0",
    "html2text>=2024.2.26",
]

[project.scripts]
nerve = "nerve.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nerve"]

[tool.hatch.build.targets.wheel.force-include]
"nerve/templates" = "nerve/templates"
